Merge to main brainch

+++ apps/steam-auth/src/app/auth/abstract/abstract.login.ts
@@ -1,14 +1,15 @@
 import * as puppeteer from 'puppeteer';
 import { PuppeteerService } from '../../puppeteer/puppeteer.service';
-import { BaseLoginRequest } from '../../../../../../libs/nestjs/src/lib/shared/dto/auth/BaseLoginRequest';
 import { TASK_NAMES } from '../../shared/enums';
+import { LoginRequest } from '../../shared/dto/BaseLoginRequest';
+import { Injectable } from '@nestjs/common';
 
-interface LoginOptions {
+export interface LoginOptions {
     closePage?: boolean
 }
 
 interface ExecuteParams<T, K> {
-    controllerCallback: (page: puppeteer.Page, parsedBody: T) => Promise<K>,
+    controllerCallback: (page: puppeteer.Page, parsedBody: T, options?: LoginOptions) => Promise<K>,
     parsedBody: T,
     taskName: TASK_NAMES,
     loadCookiesFn?: (page: puppeteer.Page, username: string) => Promise<void>,
@@ -17,11 +18,11 @@ interface ExecuteParams<T, K> {
 }
 
 
-
+@Injectable()
 export class AbstractLogin {
     constructor(private readonly puppeteerService: PuppeteerService){}
 
-    async execute<T extends BaseLoginRequest, K>(
+    async execute<T extends LoginRequest, K>(
         {
             controllerCallback,
             parsedBody,
@@ -31,11 +32,11 @@ export class AbstractLogin {
             options
         }: ExecuteParams<T, K>
     ) {
-        if(options.closePage)
-            this.puppeteerService.deleteContext(parsedBody.username);
+        if(options && options.closePage)
+            await this.puppeteerService.deleteContext(parsedBody.username);
 
         const taskFn = (page: puppeteer.Page): Promise<K> =>
-            controllerCallback(page, parsedBody);
+            controllerCallback(page, parsedBody, options);
 
         return await this.puppeteerService.executePuppeteerTask(
             parsedBody.username,
